Added javadocs to Number class and its extensions.
