+++ src/model/ControlUnitImpl.java
@@ -10,7 +10,7 @@ import view.SimulatorWindow;
 /**
  * 
  * @author
- * @version 2.0
+ * @version 2.9
  *
  */
 public class ControlUnitImpl implements ControlUnit {
